+++ examples/SimpleBotShield/ping.js
@@ -8,7 +8,11 @@ var board = new five.Board(opts);
 board.on("ready", function() {
 
     //Create new Ping and show distance on change
-    var ping = new five.Ping(8);
+    var ping = new five.Proximity({
+        pin: 8,
+        controller: "HCSR04",
+        freq: 200,
+    });
 
     ping.on("change", function( err, value ) {